FIXUP COMMIT: This is a large commit, and represents a majority of the work required to migrate logic into the context struct. Progress bar messages are now set in the context's wrapper functions, which call down into separate functions responsible for the actual implementation of an action. This commit will not compile if `cargo test` is run, there is still refactoring work to do to apply these changes to the test bench.
