+++ src/manifest/mod.rs
@@ -6,7 +6,6 @@ use emoji;
 use error::Error;
 use serde_json;
 use toml;
-use PBAR;
 
 mod cargo;
 mod npm;
@@ -22,43 +21,6 @@ pub fn read_cargo_toml(path: &str) -> Result<CargoManifest, Error> {
     cargo_file.read_to_string(&mut cargo_contents)?;
 
     Ok(toml::from_str(&cargo_contents)?)
-}
-
-/// Generate a package.json file inside in `./pkg`.
-pub fn write_package_json(path: &str, scope: &Option<String>) -> Result<(), Error> {
-    let step = format!(
-        "{} {}Writing a package.json...",
-        style("[4/7]").bold().dim(),
-        emoji::MEMO
-    );
-
-    let warn_fmt = |field| {
-        format!(
-            "Field {} is missing from Cargo.toml. It is not necessary, but recommended",
-            field
-        )
-    };
-
-    let pb = PBAR.message(&step);
-    let pkg_file_path = format!("{}/pkg/package.json", path);
-    let mut pkg_file = File::create(pkg_file_path)?;
-    let manifest = read_cargo_toml(path)?;
-    let npm_data = NpmPackage::new(&manifest, scope);
-
-    if npm_data.description.is_none() {
-        PBAR.warn(&warn_fmt("description"));
-    }
-    if npm_data.repository.is_none() {
-        PBAR.warn(&warn_fmt("repository"));
-    }
-    if npm_data.license.is_none() {
-        PBAR.warn(&warn_fmt("license"));
-    }
-
-    let npm_json = serde_json::to_string_pretty(&npm_data)?;
-    pkg_file.write_all(npm_json.as_bytes())?;
-    pb.finish();
-    Ok(())
 }
 
 pub fn get_crate_name(path: &str) -> Result<String, Error> {
